+++ gcs/src/main/java/com/google/cloud/hadoop/fs/gcs/GoogleHadoopFileSystemConfiguration.java
@@ -32,7 +32,6 @@ import com.google.cloud.hadoop.gcsio.GoogleCloudStorageReadOptions;
 import com.google.cloud.hadoop.gcsio.GoogleCloudStorageReadOptions.Fadvise;
 import com.google.cloud.hadoop.gcsio.PerformanceCachingGoogleCloudStorageOptions;
 import com.google.cloud.hadoop.gcsio.authorization.AuthorizationHandler;
-import com.google.cloud.hadoop.gcsio.authorization.AuthorizationMode;
 import com.google.cloud.hadoop.gcsio.cooplock.CooperativeLockingOptions;
 import com.google.cloud.hadoop.util.AsyncWriteChannelOptions;
 import com.google.cloud.hadoop.util.AsyncWriteChannelOptions.PipeType;
@@ -420,10 +419,6 @@ public class GoogleHadoopFileSystemConfiguration {
           new HadoopConfigurationProperty<>(
               "fs.gs.authorization.handler.properties.", ImmutableMap.of());
 
-  /** Configuration for the mode to generate AccessToken. */
-  public static final HadoopConfigurationProperty<AuthorizationMode> GCS_AUTHORIZATION_MODE =
-      new HadoopConfigurationProperty<>("fs.gs.authorization.mode", AuthorizationMode.BASIC);
-
   // TODO(b/120887495): This @VisibleForTesting annotation was being ignored by prod code.
   // Please check that removing it is correct, and remove this comment along with it.
   // @VisibleForTesting
@@ -484,8 +479,7 @@ public class GoogleHadoopFileSystemConfiguration {
             GCS_AUTHORIZATION_HANDLER_IMPL.get(
                 config, (k, d) -> config.getClass(k, d, AuthorizationHandler.class)))
         .setAuthorizationHandlerProperties(
-            GCS_AUTHORIZATION_HANDLER_PROPERTIES_PREFIX.getPropsWithPrefix(config))
-        .setAuthorizationMode(GCS_AUTHORIZATION_MODE.get(config, config::getEnum));
+            GCS_AUTHORIZATION_HANDLER_PROPERTIES_PREFIX.getPropsWithPrefix(config));
   }
 
   private static PerformanceCachingGoogleCloudStorageOptions getPerformanceCachingOptions(