+++ gcsio/src/main/java/com/google/cloud/hadoop/gcsio/GoogleCloudStorageOptions.java
@@ -20,7 +20,6 @@ import static com.google.common.base.Strings.isNullOrEmpty;
 import com.google.api.services.storage.Storage;
 import com.google.auto.value.AutoValue;
 import com.google.cloud.hadoop.gcsio.authorization.AuthorizationHandler;
-import com.google.cloud.hadoop.gcsio.authorization.AuthorizationMode;
 import com.google.cloud.hadoop.gcsio.cooplock.CooperativeLockingOptions;
 import com.google.cloud.hadoop.util.AsyncWriteChannelOptions;
 import com.google.cloud.hadoop.util.HttpTransportFactory;
@@ -88,9 +87,6 @@ public abstract class GoogleCloudStorageOptions {
   public static final Class<? extends AuthorizationHandler>
       AUTHORIZATION_HANDLER_IMPL_CLASS_DEFAULT = null;
 
-  /** Default setting for generating new access token per request */
-  public static AuthorizationMode AUTHORIZATION_MODE_DEFAULT = AuthorizationMode.BASIC;
-
   /** Default properties for authorization handler. */
   public static final Map<String, String> AUTHORIZATION_HANDLER_PROPERTIES_DEFAULT =
       ImmutableMap.of();
@@ -120,7 +116,6 @@ public abstract class GoogleCloudStorageOptions {
         .setCooperativeLockingOptions(CooperativeLockingOptions.DEFAULT)
         .setHttpRequestHeaders(HTTP_REQUEST_HEADERS_DEFAULT)
         .setAuthorizationHandlerImplClass(AUTHORIZATION_HANDLER_IMPL_CLASS_DEFAULT)
-        .setAuthorizationMode(AUTHORIZATION_MODE_DEFAULT)
         .setAuthorizationHandlerProperties(AUTHORIZATION_HANDLER_PROPERTIES_DEFAULT);
   }
 
@@ -190,16 +185,8 @@ public abstract class GoogleCloudStorageOptions {
   @Nullable
   public abstract RedactedString getEncryptionKeyHash();
 
-  @Nullable
-  public abstract Class<? extends AuthorizationHandler> getAuthorizationHandlerImplClass();
-
-  public abstract Map<String, String> getAuthorizationHandlerProperties();
-
-  public abstract AuthorizationMode getAuthorizationMode();
-
   public RetryHttpInitializerOptions toRetryHttpInitializerOptions() {
     return RetryHttpInitializerOptions.builder()
-        // When access tokens are refreshed per request, low level HTTP retry should be disabled.
         .setDefaultUserAgent(getAppName())
         .setHttpHeaders(getHttpRequestHeaders())
         .setMaxRequestRetries(getMaxHttpRequestRetries())
@@ -207,6 +194,11 @@ public abstract class GoogleCloudStorageOptions {
         .setReadTimeout(Duration.ofMillis(getHttpRequestReadTimeout()))
         .build();
   }
+
+  @Nullable
+  public abstract Class<? extends AuthorizationHandler> getAuthorizationHandlerImplClass();
+
+  public abstract Map<String, String> getAuthorizationHandlerProperties();
 
   public void throwIfNotValid() {
     checkArgument(!isNullOrEmpty(getAppName()), "appName must not be null or empty");
@@ -277,8 +269,6 @@ public abstract class GoogleCloudStorageOptions {
 
     public abstract Builder setAuthorizationHandlerImplClass(
         Class<? extends AuthorizationHandler> authorizationHandlerImpl);
-
-    public abstract Builder setAuthorizationMode(AuthorizationMode authorizationMode);
 
     public abstract Builder setAuthorizationHandlerProperties(Map<String, String> properties);